fix model props type

+++ schemas/models.ts
@@ -15,8 +15,8 @@ export const modelSchema = z.object({
         .int()
         .min(Categories.Roof, 'Category ID must be at least 1')
         .max(Categories.Foundation, 'Category ID must be at most 5'),
-    width: z.number().positive(),
-    height: z.number().positive(),
+    width: z.number().int().positive(),
+    height: z.number().int().positive(),
     position: z.enum(['horizontal', 'vertical']),
 })
